region_envelope_injector/requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
streamlit>=1.30
imageio>=2.31
imageio-ffmpeg>=0.4.9
